refactor: replace path-variable userId with authentication
