+++ src/main/java/com/solution/Ongi/domain/meal/dto/MealScheduleResponse.java
@@ -1,24 +1,24 @@
 package com.solution.Ongi.domain.meal.dto;
 
-import com.fasterxml.jackson.annotation.JsonFormat;
 import com.solution.Ongi.domain.meal.MealSchedule;
-import com.solution.Ongi.domain.meal.enums.MealType;
 
+import java.time.LocalDate;
 import java.time.LocalTime;
 
 public record MealScheduleResponse(
         Long scheduleId,
-        @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "HH:mm")
-        LocalTime mealTime,
-        Boolean status,
-        MealType mealType
+        LocalDate mealScheduleDate,
+        LocalTime mealScheduleTime,
+        boolean status,
+        Long mealId
 ) {
-    public static MealScheduleResponse from(MealSchedule schedule) {
+    public static MealScheduleResponse from(MealSchedule mealSchedule) {
         return new MealScheduleResponse(
-                schedule.getId(),
-                schedule.getMeal_schedule_time(),
-                schedule.isStatus(),
-                schedule.getMeal().getMeal_type()
+                mealSchedule.getId(),
+                mealSchedule.getMealScheduleDate(),
+                mealSchedule.getMealScheduleTime(),
+                mealSchedule.isStatus(),
+                mealSchedule.getMeal().getId()
         );
     }
 }
