+++ src/main/java/com/solution/Ongi/domain/meal/dto/MealScheduleWithUserResponse.java
@@ -1,11 +1,9 @@
 package com.solution.Ongi.domain.meal.dto;
 
-import com.solution.Ongi.domain.meal.MealSchedule2;
-
 import java.time.LocalDate;
 import java.time.LocalTime;
 
-public record MealSchedule2WithUserResponse(
+public record MealScheduleWithUserResponse(
         Long scheduleId,
         LocalDate scheduleDate,
         LocalTime mealScheduleTime,
