+++ src/main/java/com/solution/Ongi/domain/meal/service/MealScheduleService.java
@@ -3,27 +3,19 @@ package com.solution.Ongi.domain.meal.service;
 
 import com.solution.Ongi.domain.meal.Meal;
 import com.solution.Ongi.domain.meal.MealSchedule;
-import com.solution.Ongi.domain.meal.MealSchedule2;
 import com.solution.Ongi.domain.meal.dto.MealScheduleResponse;
 import com.solution.Ongi.domain.meal.dto.UpdateMealScheduleStatusesRequest;
 import com.solution.Ongi.domain.meal.repository.MealRepository;
 import com.solution.Ongi.domain.meal.repository.MealScheduleRepository;
+import com.solution.Ongi.domain.user.User;
 import com.solution.Ongi.domain.user.service.UserService;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
 import java.util.List;
-
-/*
- * create -> Meal 생성 시점에 동시에 생성 (all null로)-> cascade 로 커버쳐지나? o
- * 주기적 초기화 (update_meal status 만 false 로 초기화)   o
- * 오늘 날짜에 해당하는 MealSchedule 조회 서비스 (get)(메인 화면 서비스) -> 걍 띄우면 되네~ 개이득  o
- * 상태 체크 (update), count ++ -> user 멤버로 넣는게 나을듯?
- * 누락 횟수 서비스 (user service)
- * 오늘 복약 처리 변경
-*/
 
 @Service
 @Transactional
@@ -35,18 +27,26 @@ public class MealScheduleService {
     private final UserService userService;
 
     public MealSchedule createMealSchedule(Meal meal){
-        MealSchedule schedule=MealSchedule.builder()
+        MealSchedule schedule= MealSchedule.builder()
                 .meal(meal)
-                .meal_schedule_time(meal.getMeal_time())
+                .mealScheduleTime(meal.getMeal_time())
+                .mealScheduleDate(LocalDate.now())
                 .status(false)
                 .build();
 
         return mealScheduleRepository.save(schedule);
     }
 
+    //매일 자정 meal schedule 생성
+    @Scheduled(cron="0 0 0 * * *")
+    public void createDailyMealSchedule(){
+        mealRepository.findAll()
+                .forEach(this::createMealSchedule);
+    }
+
     //단일 schedule 상태 업데이트
     public void updateMealScheduleStatus(Long scheduleId,boolean newStatus){
-        MealSchedule mealSchedule=mealScheduleRepository.findById(scheduleId)
+        MealSchedule mealSchedule= mealScheduleRepository.findById(scheduleId)
                 .orElseThrow(()->new RuntimeException("스케줄 ID " + scheduleId + "를 찾을 수 없습니다."));
 
         mealSchedule.setStatus(newStatus);
@@ -60,29 +60,31 @@ public class MealScheduleService {
         }
     }
 
-    //매일 자정 status 초기화
-    //-> 매일 자정 생성
-    @Scheduled(cron = "0 0 0 * * *")
-    public void resetMealScheduleStatus(){
-//        List<MealSchedule>mealSchedules=mealScheduleRepository.findAll();
-//        for(MealSchedule mealSchedule:mealSchedules){
-//            mealSchedule.setStatus(false);
-//        }
-//        mealScheduleRepository.saveAll(mealSchedules);
-    }
-
     //오늘 meal schedule 가져오기
-    //현재 날짜에 대응하는 MealSchedule 가져와
-    public List<MealSchedule> getMealSchedulesByUserId(Long userId){
-        userService.getUserByIdOrThrow(userId);
-        return mealScheduleRepository.findByMeal_User_Id(userId);
+    //날짜에 대응하는 MealSchedule 조회
+    public List<MealSchedule> getMealSchedulesByUserId(String loginId){
+        User user=userService.getUserByLoginIdOrThrow(loginId);
+        return mealScheduleRepository.findByMeal_User_IdAndMealScheduleDate(user.getId(),LocalDate.now());
     }
 
-    /*
-     * 수정 사안
-     * 1. 조회 가능해야 -> 일대다 매핑으로
-     * 2. 스케줄러 -> 값 초기화가 아닌 생성으로
-     * 3. 추가: 해당 날짜 조회
-     * +) 사용자 이름 반환
-     */
+    //특정 날짜 meal schedule 조회
+    public List<MealScheduleResponse> getMealSchedulesByExactDate(String loginId, LocalDate date){
+        User user=userService.getUserByLoginIdOrThrow(loginId);
+    //    return mealScheduleRepository.findByMeal_User_IdAndMealScheduleDate(user.getId(),date);
+        return mealScheduleRepository
+                .findByMeal_User_IdAndMealScheduleDate(user.getId(),date)
+                .stream()
+                .map(MealScheduleResponse::from)
+                .toList();
+    }
+
+    //생성 시점이 특정 기한 내에 속하는 meal schedule 조회
+    public List<MealScheduleResponse> getMealSchedulesByDate(String loginId, LocalDate startDate, LocalDate endDate){
+        User user=userService.getUserByLoginIdOrThrow(loginId);
+        return mealScheduleRepository
+                .findByMeal_User_IdAndMealScheduleDateBetween(user.getId(), startDate, endDate)
+                .stream()
+                .map(MealScheduleResponse::from)
+                .toList();
+    }
 }
